Refactor renamed files and imports to be more inline with PEP8
