+++ backend/api/views/detection_detail_idref_query_create.py
